requirements.txt
django==5.1.1
djangorestframework==3.15
python-dotenv==1.0.1